Adapted TimeEntryHolder to hold time entry groups.
