+++ Phoebe/Data/Views/LogTimeEntriesView.cs
@@ -595,212 +595,6 @@ namespace Toggl.Phoebe.Data.Views
             }
         }
 
-        public class TimeEntryHolder
-        {
-            private TimeEntryData timeEntry;
-            private ProjectData project;
-            private ClientData client;
-            private TaskData task;
-            private int numberOfTags;
-
-            public TimeEntryHolder (TimeEntryData timeEntry)
-            {
-                this.timeEntry = new TimeEntryData (timeEntry);
-                project = new ProjectData ();
-                client = new ClientData ();
-                task = new TaskData ();
-            }
-
-            public TimeEntryData TimeEntryData
-            {
-                get { return timeEntry; }
-            }
-
-            public ProjectData ProjectData
-            {
-                get { return project; }
-            }
-
-            public ClientData ClientData
-            {
-                get { return client; }
-            }
-
-            public TaskData TaskData
-            {
-                get { return task; }
-            }
-
-            public Guid Id
-            {
-                get { return TimeEntryData.Id; }
-            }
-
-            public int NumberOfTags
-            {
-                get { return numberOfTags; }
-            }
-
-            public string ProjectName
-            {
-                get {
-                    return project.Name;
-                }
-            }
-
-            public string ClientName
-            {
-                get {
-                    return client.Name;
-                }
-            }
-
-            public string TaskName
-            {
-                get {
-                    return task.Name;
-                }
-            }
-
-            public string Description
-            {
-                get {
-                    return timeEntry.Description;
-                }
-            }
-
-            public int Color
-            {
-                get {
-                    return (project.Id != Guid.Empty) ? project.Color : -1;
-                }
-            }
-
-            public TimeEntryState State
-            {
-                get {
-                    return timeEntry.State;
-                }
-            }
-
-            public bool IsBillable
-            {
-                get {
-                    return timeEntry.IsBillable;
-                }
-            }
-
-            public async Task LoadAsync ()
-            {
-                numberOfTags = 0;
-
-                if (timeEntry.ProjectId.HasValue) {
-                    project = await GetProjectDataAsync (timeEntry.ProjectId.Value);
-                    if (project.ClientId.HasValue) {
-                        client = await GetClientDataAsync (project.ClientId.Value);
-                    }
-                }
-
-                if (timeEntry.TaskId.HasValue) {
-                    task = await GetTaskDataAsync (timeEntry.TaskId.Value);
-                }
-
-                numberOfTags = await GetNumberOfTagsAsync (timeEntry.Id);
-            }
-
-            public async Task UpdateAsync (TimeEntryData data)
-            {
-                timeEntry = new TimeEntryData (data);
-                project = await UpdateProject (data, ProjectData);
-                client = await UpdateClient (project, ClientData);
-                task = await UpdateTask (data, TaskData);
-                numberOfTags = await GetNumberOfTagsAsync (data.Id);
-            }
-
-            private async Task<ProjectData> GetProjectDataAsync (Guid projectGuid)
-            {
-                var store = ServiceContainer.Resolve<IDataStore> ();
-                var projectList = await store.Table<ProjectData> ()
-                                  .Take (1).QueryAsync (m => m.Id == projectGuid);
-                return projectList.First ();
-            }
-
-            private async Task<TaskData> GetTaskDataAsync (Guid taskId)
-            {
-                var store = ServiceContainer.Resolve<IDataStore> ();
-                var taskList = await store.Table<TaskData> ()
-                               .Take (1).QueryAsync (m => m.Id == taskId);
-                return taskList.First ();
-            }
-
-            private async Task<ClientData> GetClientDataAsync (Guid clientId)
-            {
-                var store = ServiceContainer.Resolve<IDataStore> ();
-                var clientList = await store.Table<ClientData> ()
-                                 .Take (1).QueryAsync (m => m.Id == clientId);
-                return clientList.First ();
-            }
-
-            private Task<int> GetNumberOfTagsAsync (Guid timeEntryGuid)
-            {
-                var store = ServiceContainer.Resolve<IDataStore> ();
-                return store.Table<TimeEntryTagData>()
-                       .Where (t => t.TimeEntryId == timeEntryGuid)
-                       .CountAsync ();
-            }
-
-            private async Task<ProjectData> UpdateProject (TimeEntryData newTimeEntry, ProjectData oldProjectData)
-            {
-                if (!newTimeEntry.ProjectId.HasValue) {
-                    return new ProjectData ();
-                }
-
-                if (oldProjectData.Id == Guid.Empty && newTimeEntry.ProjectId.HasValue) {
-                    return await GetProjectDataAsync (newTimeEntry.ProjectId.Value);
-                }
-
-                if (newTimeEntry.ProjectId.Value != oldProjectData.Id) {
-                    return await GetProjectDataAsync (newTimeEntry.ProjectId.Value);
-                }
-
-                return oldProjectData;
-            }
-
-            private async Task<ClientData> UpdateClient (ProjectData projectData, ClientData oldClientData)
-            {
-                if (!projectData.ClientId.HasValue) {
-                    return new ClientData ();
-                }
-
-                if (oldClientData == null && projectData.ClientId.HasValue) {
-                    return await GetClientDataAsync (projectData.ClientId.Value);
-                }
-
-                if (projectData.ClientId.Value != oldClientData.Id) {
-                    return await GetClientDataAsync (projectData.ClientId.Value);
-                }
-
-                return oldClientData;
-            }
-
-            private async Task<TaskData> UpdateTask (TimeEntryData newTimeEntry, TaskData oldTaskData)
-            {
-                if (!newTimeEntry.TaskId.HasValue) {
-                    return new TaskData ();
-                }
-
-                if (oldTaskData == null && newTimeEntry.TaskId.HasValue) {
-                    return await GetTaskDataAsync (newTimeEntry.TaskId.Value);
-                }
-
-                if (newTimeEntry.TaskId.Value != oldTaskData.Id) {
-                    return await GetTaskDataAsync (newTimeEntry.TaskId.Value);
-                }
-
-                return oldTaskData;
-            }
-        }
-
         private enum UpdateMode {
             Immediate,
             Batch,
